app/requirements.txt
fastapi
uvicorn[standard]
jinja2
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
